+++ src/lib/config/browser.ts
@@ -1,0 +1,46 @@
+import type { BrowserDetails } from "@/lib/types/types";
+
+export const browserDetails: BrowserDetails[] = [
+  {
+    name: "Mozilla Firefox",
+    short: "firefox",
+    href: "https://addons.mozilla.org/en-US/firefox/addon/recommen-do/",
+    ariaLabel: "Add extension to Firefox",
+    description: "Add extension to Firefox",
+  },
+  {
+    name: "Google Chrome",
+    short: "chrome",
+    href: "https://chrome.google.com/webstore/detail/ai-recommendations-for-sh/obfbgdconmhiolihlenkaopigkpeblne",
+    ariaLabel: "Get extension for Chrome",
+    description: "Get extension for Chrome",
+  },
+  {
+    name: "Microsoft Edge",
+    short: "edge",
+    href: "https://microsoftedge.microsoft.com/addons/detail/recommendo/pdahoiahedcdggmdopbgefclpcpeiioc",
+    ariaLabel: "Get extension for Edge",
+    description: "Get extension for Edge",
+  },
+  {
+    name: "Brave",
+    short: "brave",
+    href: "https://chrome.google.com/webstore/detail/ai-recommendations-for-sh/obfbgdconmhiolihlenkaopigkpeblne",
+    ariaLabel: "Get extension for Brave",
+    description: "Get extension for Brave",
+  },
+  {
+    name: "Opera",
+    short: "opera",
+    href: "https://chrome.google.com/webstore/detail/ai-recommendations-for-sh/obfbgdconmhiolihlenkaopigkpeblne",
+    ariaLabel: "Get extension for Opera",
+    description: "Get extension for Opera",
+  },
+  {
+    name: "Vivaldi",
+    short: "vivaldi",
+    href: "https://chrome.google.com/webstore/detail/ai-recommendations-for-sh/obfbgdconmhiolihlenkaopigkpeblne",
+    ariaLabel: "Get extension for Vivaldi",
+    description: "Get extension for Vivaldi",
+  },
+];